Initial version of the dashboard tabs
